backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
pydantic-settings>=2.6
httpx>=0.27
apscheduler>=3.10
